Prevent excessive frame events when we ignore a frame
In wlroots, if the compositor doesn't swap the output's buffers in
response to a frame event then wlroots would emit a frame event whenever
the Wayland event loop becomes idle, and do so repeatedly, because the
idle event source is only removed when the buffers are swapped.

A proper fix for this is to make wlroots request frames from the
backend, but in the interim this patch disables the idle event source if
we're ignoring the frame event.

This prevents sway from throttling the CPU when mpv displays album art
when mpv is using the wayland backend.

This also adds a missing call to pixman_region32_fini.